A FIX ON RX1 AND RX2 RECEIVE WINDOW DURATION
This fix subsitutes the constant receive window duration for Rx1 and Rx2 windows with a duration that corsponds the particular spreading factror and bandwidth they are configured with. This will improve the power consumption calculation as it will switch the transceiver to standby mode for the duration of the time that matches the receiver configuration. It also will increase the overlap between this windows and other transmissions near the recieve window.