Do not update FCnt for re-transmissions
This commit fixes issue #68. In addition to remove the line that updates
the FCnt, this commit also introduces a new example to test the behavior
of re-transmissions, along with some changes in the NetworkScheduler
to remove its reliance on the frame counter to determine whether a
packet was received by another gateway already.