LoraTag: Add additional getter/setter
This adds more getter/setter to also be able to store the channel
bandwidth, the signal to noise ration, and the sync word.

Actual users of the new API need to be added as follow-up. For the
bandwidth 125 kHz and for the sync word 0x34 (LoRaWAN) are used as
default, which matches expectations of the model. Actual use cases for
this API extension need to be added as follow up.