LoraPhy, LoraChannel: simulate sync word behavior
This adds a simple sync word model: A PHY is configured with a given
sync word and will use that in the preamble of outgoing packets. For
incoming packets, the PHY will only lock onto the transmission if the
sync word of the incoming packet matches the configured sync word.